• composition.py   - first draft using PNLProgress and _report_output
